backend/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.9.2
mangum==0.17.0   # only needed if you deploy to AWS Lambda via API Gateway
python-dotenv==1.0.1
python-multipart==0.0.9
PyPDF2==3.0.1
aiofiles==24.1.0
